оптимизировал simpleClient
